apps/python_app/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10